+++ CHANGELOG.md
@@ -12,6 +12,7 @@
     * `stubbles\lang\reflect\ReflectionParameter::annotation()`
     * `stubbles\lang\reflect\ReflectionProperty::annotation()`
   * Deprecated `stubbles\lang\reflect\Reflection*::getAnnotation()`, use `stubbles\lang\reflect\Reflection*::annotation()` instead
+  * Parsing `null` with any of the `stubbles\lang\Parse` methods will now always return null.
 
 
 ### Other changes